improvement(zim-to-website): use the original wiki homepage for redirect rather than kiwix
