+++ zim-to-website/src/index.ts
@@ -1,5 +1,6 @@
 import { Command, flags } from '@oclif/command'
 import { zimToWebsite } from './zim-to-website'
+import { Options } from './domain'
 
 class ZimToWebsite extends Command {
   static description = 'describe the command here'
@@ -14,12 +15,13 @@ class ZimToWebsite extends Command {
   async run() {
     const { args } = this.parse(ZimToWebsite)
 
-    const options = {
+    const options: Options = {
       unpackedZimDir: args.unpackedZimDir,
       host: 'tr.wikipedia-on-ipfs.org',
       ipnsHash: 'QmVH1VzGBydSfmNG7rmdDjAeBZ71UVeEahVbNpFQtwZK8W',
       zimFile: 'wikipedia_tr_all_maxi_2019-12.zim',
-      mainPage: 'Kullanıcı:The_other_Kiwix_guy/Landing'
+      kiwixMainPage: 'Kullanıcı:The_other_Kiwix_guy/Landing',
+      mainPage: 'Anasayfa.html'
     }
 
     await zimToWebsite(options)
